Add typedefs
Related To: #25 

### What does this do?
The goal of this PR is to add accurate type definitions to MathBox by converting a small portion of the source code to typescript.

Here's the approach I've taken: 

**Nodes, Traits, Properties:** MathBox has 64 node variants (axis, area, cartesian, etc). These are the APIs we usually interact with. Each Node has many properties. For example, `<line>` nodes have [20 properties](https://gitgud.io/unconed/mathbox/-/blob/master/docs/primitives.md#draw/line) (size, start, etc). 

Nodes do not directly define their properties. Instead, nodes are assigned "traits" and each trait implements a collection of properties. For example, `<line>` nodes implement the traits `["node", "object", "style", "line", "arrow", "geometry"]`. [Footnote 1], and the "arrow" trait implements properties roughly ~ `{ size: number, start: boolean, end: boolean }`.

On the master branch, the trait <--> property relationship is defined in `src/primitives/types/traits.js`... So:

1. Convert `traits.js` to typescript. This file depends on `types.js`, which we will manually type. As long as our manual typings `types.js` are correct, we can have high confidence in anything whose types are based on `traits.ts`.

2. Define interfaces for each trait. This is a bit tedious but not very hard:
    ```
    import { Traits as TraitsValue } from "./primitives/types/traits";
    
    // This interface is based on traits.ts, so we can be confident it is correct.
    interface SetTraitsArrow {
      size: Parameters<Traits["arrow"]["size"]["validate"]>[0];
      start: Parameters<Traits["arrow"]["start"]["validate"]>[0];
      end: Parameters<Traits["arrow"]["end"]["validate"]>[0];
    }
    ``` 
3. Define interfaces for `< line >` node properties:
    ```ts 
    // Each of these is based on `traits.ts`, so we can have high confidence 
    export interface LineProps
      extends SetTraitsNode,
        SetTraitsObject,
        SetTraitsStyle,
        SetTraitsLine,
        SetTraitsArrow,
        SetTraitsGeometry {}
    ```

### To-do:
The above approach is working very well. Remaining TODOs:
- [ ] finish manual typings for `types.js`
- [ ] add more tests to `types.spec.ts`. Since that test file is written in typescript, if its tests pass, we can be confident in my manual typings.

### Also, documentation!
Run `npm run typedoc` to build docs with [typedoc](https://typedoc.org/) and open `docs/index.html`. Note: The index.html file is build from the README. Click "Exports" to view the API documentation.

The resulting documentation is IMO pretty good. It's not as compact as the existing `primitives.md` reference, but it is more correct and pretty navigable. IMO it would be good to replace the existing API-reference style documentation with something like this, but that's another topic.